requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
yt-dlp
openai>=2.0.0
google-generativeai==0.8.0
pydantic-settings==2.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
imageio-ffmpeg>=0.5.1
